group_manager_bot/requirements.txt
# Rubika Group Manager Bot Requirements

# Core library (rubpy should be installed from the parent project)
# rubpy

# Environment variable handling
python-dotenv>=0.19.0

# Logging and monitoring
python-logging>=1.4.1

# Optional enhancements
aiohttp>=3.8.0
